feat(core): expose browser-safe tool identity contract
